[UPD] cg: separate class ExtractorModuleFile from extrator_module.py
